requirements.txt
accelerate==0.29.3
appnope==0.1.4
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
diffusers==0.27.2
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.13.4
fsspec==2024.3.1
huggingface-hub==0.22.2
idna==3.7
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
pyzmq==26.0.2
regex==2024.4.16
requests==2.31.0
safetensors==0.4.3
six==1.16.0
stack-data==0.6.3
sympy==1.12
tokenizers==0.19.1
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformers==4.40.1
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
zipp==3.18.1
